+++ tests/ChromeDriverConnectionTest.php
@@ -2,12 +2,10 @@
 
 namespace DMore\ChromeDriverTests;
 
-use Behat\Mink\Exception\ElementNotFoundException;
-use Behat\Mink\Exception\DriverException;
-use DMore\ChromeDriver\ChromeBrowser as Browser;
+use DMore\ChromeDriver\ChromeBrowser;
 use DMore\ChromeDriver\ChromeDriver;
 use DMore\ChromeDriver\HttpClient;
-use PHPUnit\Framework\TestCase;
+use WebSocket\TimeoutException;
 
 /**
  * Note that the majority of driver test coverage is provided via minkphp/driver-testsuite.
@@ -16,10 +14,43 @@ use PHPUnit\Framework\TestCase;
  */
 class ChromeDriverConnectionTest extends ChromeDriverTestBase
 {
-   /**
-     * @throws DriverException
+    /**
+     * Unable to connect to nonsense ChromeDriver URL.
      */
-    public function testInformativeExceptionIfChromeConnectionFailed()
+    public function testRuntimeExceptionIfNotConnected()
+    {
+        $nonWorkingUrl = 'http://localhost:12345';
+        $this->driver = new ChromeDriver($nonWorkingUrl, null, 'about:blank', []);
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Could not fetch version information from http://localhost:12345/json/version.');
+        $this->driver->visit('about:blank');
+        // Content read is necessary to trigger timeout.
+        $this->driver->getContent();
+    }
+
+    /**
+     * JS confirm() will lead the browser to time out.
+     */
+    public function testTimeoutExceptionIfResponseBlocked()
+    {
+        // We don't want to wait the default 10s to time out.
+        $options = [
+            'socketTimeout' => 1,
+        ];
+        $this->driver = new ChromeDriver('http://localhost:9222', null, 'about:blank', $options);
+        $this->expectException(TimeoutException::class);
+        $this->expectExceptionMessage('Client read timeout');
+        $script = "confirm('Is the browser blocked? (yes, it is)');";
+        $this->driver->visit('about:blank');
+        $this->driver->evaluateScript($script);
+        // Content read is necessary to trigger timeout.
+        $this->driver->getContent();
+    }
+
+    /**
+     *
+     */
+    public function testRuntimeExceptionIfClientConnectionFails()
     {
         $client = $this->getMockBuilder(HttpClient::class)
             ->disableOriginalConstructor()
@@ -30,10 +61,10 @@ class ChromeDriverConnectionTest extends ChromeDriverTestBase
             ->willReturn('Error Happened!');
 
         $this->expectException(\RuntimeException::class);
-// Test that chromium response is included in exception message.
+        // Test that chromium response is included in exception message.
         $this->expectExceptionMessageMatches('/Error Happened!/');
 
-        $browser = new Browser('http://localhost:9222');
+        $browser = new ChromeBrowser('http://localhost:9222');
         $browser->setHttpClient($client);
         $browser->setHttpUri('http://localhost:9222');
         $browser->start();